housing-dashboard-backup/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.2
python-multipart==0.0.6
starlette==0.27.0
